[Spesh] Switch order of insert
We did a insert-before-current, which we then had to fix up because
the annotations would be on the wrong operator; then we'd have to call
both the optimize_istrue_isfalse and optimize_iffy (because the
optimizer doesn't process the instructions 'behind' it). And the more
I thought about it, the sillier that really is. If we stick to the
convention that whenever we add an instruction we push it forward, we
will never have such problems, and moreover, the behaviour of the
optimizer as a whole is far more understandable.