[Spesh] Add optimize_unbox
The original optimize_iffy had a special loop for unboxing boxed
integers (and other things). I've taken that out and generalized it
for 'unbox' operations. But as far as I've been able to determine,
that isn't actually effective (the optimization is never
applied). Maybe that is because rakudo outputs better code than
before? Anyway, I've kept it but disabled it.

I think we can improve on it if we allocate a (unique) register for a
'set' between the original boxed value and the unbox. (This has to be
unique because we insert the 'set' backwards, so otherwise we might
use a register that we clobber inbetween).