profiler: make sure returning from inlined frames gets logged
it was possible for an optimized return (the kind that works
without building an actual exception object) to leave inlined
portions of code without signalling to the profiler that a
frame was in fact left.

This led to call graphs that had copies of the same routine
nested very deeply inside of itself, making the call graph
rather useless and the profile files rather gigantic.