Speed up index 50% for flat haystack and diff type needle
Convert the needle to match the haystack's data type when encountering a
flat haystack (very common since Perl 6 flattens the haystack during
regex).

Also fix a problem in the 32 bit memmem loop. The loop runs memmem again
if it finds a match not on a 32 bit boundary (since memmem is 8bit
based). It then sets the start position to the returned position and
runs it again. This could enter an infinite loop, since it did not set
the start position to *after* the previous match.

This can likely only be triggered by the needle being a \0 grapheme
and probably couldn't have been triggered at all before these changes
which convert the needle's type to match the haystack. Since a \0
grapheme needle fits in 8 bits and is a single grapheme.