Use Configure.pl to generate vectorization defines
MSVC doesn't support the C99 feature that lets you declare a pragma
behind a #define. So instead use Configure.pl to add it for clang.

We set this for clang because it will give us a warning if it fails to
vectorize the code. Without it, it would vectorize, but not warn if
vectorization had failed. GCC doesn't seem to have a setting which warns
like this, so we don't set anything. The only gcc flag is
pragma GCC ivdep, but this implies that we guarentee the code won't have
improper side effects. It's probably safer not to set this if it doesn't
benefit us otherwise.