Fix fixing of wvals during inline
Whenever we inline a frame from one MVMCompUnit into a frame
from another MVMCompUnit we have to fix up several ops and operands
that are references based on "the current CU". One of those cases
was the wval operations.

Until recently, we would just grab the wval during spesh time and
stash it away in a spes slot, but if the wval hadn't been requested
before that would cause a deserialization cycle to be run, causing
objects to be allocated. That would cause a panic in spesh, because
it is not built to cooperate with GC.

A recent attempt to fix this involved storing information about what
wval was going to cause deserialization and aborting spesh of the
frame in question, then deserializing the object outside of the
complicated spesh code paths, and retrying the spesh up to 3 times.

This caused correctness issues and crashes, however.

This commit uses a different approach: Wval ops that reference
objects that haven't yet been deserialized get rewritten into an
op that stashes the SC away in a spesh slot instead of the index
into the current CU's list of dependent SCs.

It does a very similar thing for strings, but since deserializing
strings can't cause GC (since we allocate them in gen2 immediately)
we don't have to fix every occurence of string literal operands.