Fix memory corruption when running with FSA_SIZE_DEBUG set
dbg->memory is not a real pointer, it's just the placeholder for the actual
memory area (that's preceeded by the allocated size). Also we don't have to
offset (char *) by the size field as the allocator will already return the
pointer to ->memory, so *p is already the start of the actual data.

Thanks to dogbert17++ for pointing this out!