On Linux and BSD's, fall back to /dev/urandom if call fails
This is mostly important on Linux since if MoarVM was compiled on a
kernel <= 3.17 and then the user runs MoarVM on kernel without the
getrandom call, the syscall wrapper will gently return a false return
value. If this happens try /dev/urandom.

On the BSD's, they seem to guarantee the call will succeed, though
MacOS's documentation does not specify information about this.