Fix small integer return values of libffi callbacks
Apparently at least on some architectures libffi allocates a full ffi_arg
for a callbacks return value and this has to be written in full. Otherwise
we may end up with stale data in the high bytes.
libffi docs are vague about this, but do say that "rvalue must point to
storage that is sizeof(ffi_arg) or larger for non-floating point types.
For smaller-sized return value types, the ffi_arg or ffi_sarg integral type
must be used to hold the return value."
ffi_arg seems to be at least as large as a long, so use it for everything
smaller or equal to long for now. This matches behavior tested by a Debian
patch.
May fix rakudo GH #1663