For NFAs with many codepoints, use binary search
So far, we've chugged through all of the codepoints, one by one. This
is suboptimal, since the largest ones generated by the Perl 6 grammar
can have up to ~150 possibilities, which is a lot to linearly search
through. Instead, sort them once and then do a binary search.

This shaves 1% off the total instructions spent parsing a sizable
Perl 6 program, thanks to nqp_nfa_run accounting for only 87% of the
instructions that it did prior to this change.