Ensure memory blocks in the nursery are aligned where necessary
Some architectures are really picky about memory alignment (e.g. armhf for 64
bit floats). We already ensure alignment of fields in structs but if we
first allocate a memory block with uneven size, the next memory block
can end up unaligned and thus all field alignments are off. So round up
the block size. Thanks to robertle++ who debugged this and wrote the
initial version of the patch.