[JIT] Initial stack walker to find current position in frame
Currently we update the JIT entry label on every basic block entry.
This is so that the interpreter can figure out in which (possibly
inlined) frame we are (so we can look up the correct dyanamic lexical),
which exception handler is active, etc. However, this is actually
redundant, because:

- If this is the currently active JIT frame, current position within the
  frame is stored on the stack as the return address (possibly several
  frames deep), and
- If it *isn't* the current frame, we need to store the jit entry label
  anyway to support trampolining when we return to it (overwriting
  whatever was written in the beginning of the basic block).

So if we can just look up the return address from the stack when we're
looking at the current frame (which is presumably uncommon) we can avoid
a memory store on every entry of a basic block (which is presumably very
common).