[JIT] Improve stack walker to use top-of-stack
It is impractical to know how many frames we can walk back safely, and
we want to fail safely for the case that we are wrong (i.e. jit code
frame is not on stack),

Force -fno-omit-frame-pointer if the JIT is active. GNU make
automatically uses on the correct invocation of 'as', so that helps, but
I'm not sure how I'm going to configure MSVC just yet (uses ml64).