[JIT] Remove redundant dynamic label assignments
Now that we can look up the current position in the JIT frame via stack
walking, it is no longer required to load the jit entry label on the
start and end of frame handlers and at the start of every basic block.

As basic block entry is quite a bit more frequent than looking up
dynamic lexicals or throwing exceptions, I hope this should provide a
nice speedup.