Rename ci->base_code to ci->first_code to avoid confusion
We used to assume that the first codepoint of a synthetic was the base
codepoint. Back then there were no "Prepend" Unicode codepoints so this
assumption held. For a while we have supported Prepend codepoints, so
that is fine. Rename the variable so its name is accurate to its function.