profiler: fix optional parameter allocations
these ops jump to a label if the parameter was
present, but the allocation logging op was inserted
right after the op. That meant that whenever something
was allocated by these ops, i.e. due to autoboxing,
the logging was skipped over.

Instead of that, the logging ops now go in the right
place.