Implement a optimized memmem 32 bit search 4-15x faster
This is a modification of the memmem used in FreeBSD to allow us to quickly
search 32 bit strings. This is much more efficient than searching by byte
since we can skip every 4 bytes instead of every 1 byte, as well as the fact
that in a 32 bit integer, some of the bytes will be empty, making it even
less efficient.

The only caveat is the table uses a modulus so it can only jump to the next
codepoint of the same modulus. So it can only jump forward to the next
codepoint of the same modulus. This is generally not too much of a
problem and performs better than the previous 8 bit memmem for 32 bit
strings in all possible cases.