Implement memmem_two_uint32, up to 45% faster index of short 32bitstr
In cases where the first grapheme is found throughout the haystack this
can speed things up by indexing by two graphemes at the same time. To do
this we index two graphemes at a time which prevents us from finding
needles of short strings in the wrong location. Worst case it is about
the same speed on 64 bit systems. May be slower on 32-bit systems which
is why for now it is only used on 64 bit systems.
We treat the two graphemes as a 64 bit integer
and search for that, moving 32 bits at a time down the haystack, reading
each two graphemes in the haystack as a 64 bit integer and then
comparing the needle and the haystack.