[JIT] Add pointer to return address, remove stack walker
It turns out not to be necessary to search the stack for the return
address, because as we compile the JIT code frame, we know exactly where
the future return address will end up (namely, just beyond our current
stack frame). Provided we never use 'pop' and 'push', we can just store
the pointer to this address in the thread context, and read our current
position from there.

Some care is necessary to NULL this pointer (e.g. when unwinding) but
altogether this is much simpler than a stack walker, and much more
reliable too - we fully control this location. No special compiler
flags are necessary, either.