[JIT] Implement 'SAVE_RV' JIT node as RV mode
SAVE_RV would save the return value (rax) of a function to stack.
This was necessary to support the (un)boxing of strings in
nativecall. However, there already exists a mechanism for dealing with
return values in the c-call handler (the 'rv mode'). So this was
redundant, and removed with this commit.