Move uthash.h from 3rdparty to src/strings
3rdparty is really only for libraries that could potentially be replaced
by a user's local library instead of MoarVM's included library.

uthash.h is heavily modified and it is impossible for someone to use the
upstream version as our version contains custom handling for the
MVMString type. Move the file's location so it is clear that it is a
version that cannot be replaced.