Make sure cached_hash_code is never 0 to prevent hashing needlessly
Because we check if MVMString->body.cached_hash_code == 0 to tell if
we have not yet computed the hash code, ensure that hashv is never 0.

To do this we add the length of the string to hashv if and only if
hashv == 0. Since both the hashv and MVMStringIndex are both uint32,
there should never be any overflow.

The only semi-problematic case is if the string is of length 0 and
hashv is zero, though this is very very unlikely (if possible at all)
and it should be very fast to calculate the hash of a zero length
string.