Make .lines 2x faster by speeding up MVM_find_cclass for newlines
Remove the call to MVM_unicode_codepoint_has_property_value for
grapheme_is_cclass, which isn't needed since there is only one Zl
codepoint currently and it was already included in the conditional.

Add a comment as a TODO for later. We may want to have ucd2c.pl add all
Zl and Zp codepoints automatically. At the moment there are only two
codepoints that exist and it's unlikely more will be added in the
future.