Speed up MVM_string_find_cclass and cclass
Speeds up .words in rakudo a fair amount by speeding up WhiteSpace
search. Also convert code that checked for multiple General Categories
so that it checks for only one.

Add code to ucd2c.pl which allows us to check for WhiteSpace and Zl, Zp
codepoints without having to query the Unicode database to improve
performance. This creates macros which we can use in conditionals.