Handle synthetics and normalization property in string bitwise ops
Previously we could XOR, AND or OR synthetics and the result could be
explosive, creating non-existent synthetics (in addition to not doing it
by codepoint as the user would expect).

Now we iterate by codepoint and then at the end of doing the operation
for both strings we renormalize only if one of the codepoints was
greater or equal to the first significant NFC codepoint.

Do all three ops using macros so we have less repeated code for three
rarely used ops.