[JIT] No more invokish
Some MoarVM opcodes can change control flow for the interpreter, e.g by
invoking a new frame or jumping to a new ecxeption handler. The JIT
relied specially compiled 'control guards' to in order follow these
control flow changes. However, now that we actually have a pointer to
the return-address on stack, it is possible to change the 'current
position' for the JIT-compiled code, much like for interpreted
code. Thus, the control guards are no longer necessary, which should
result in a nice reduction in code size and hopefully a speedup as well.