Fix encoderep/encoderepconf returning value into wrong register
No idea why this worked before, as the return value is supposed to be written
to register 0. My guess is that it's a lucky coincidence in how the Perl 6
code gets compiled.
Oddity was introduced in commit 1bc55b7d728c8ce85599141bbdc65a30975e3582 and
then copied into the encoderepconf implementation.