Make spesh plugins work correctly with the JIT
We already put a deopt annotation on the prepargs that we are going
to delete as part of the optimization. Steal that and put it on the
guards, so that we get a viable position to use for deopt_offset. It
doesn't matter that this will not result in a very precise position,
since this is just used to answer the question of "which (if any) of
the inlines are we in". All the guards will fall within the inlines.