Only use a deopt table index once in the graph
Re-using them makes for great amounts of confusion when we take some
specialized code for inlining and reconstruct the deopt annotations. If
two annotations in the original graph shared a position in the table,
then we cannot reconstruct both. The result would be a JIT oops if we
inlined the specialization of a speshresolve that introduced more than
one guard.