Use Fibonacci hashing to determine which bucket to put hashv's into
This has an advantage over using ANDing or %ing to get the bucket number
because it uses the full bit width of the hash instead of just the last
several bits of the hashv.

We multiply the hashv by: max_hashv / phi, where max_hashv is the maximum
number that hashv can be, since ours is 32 bit that number is 2**32/phi.

Since we grow bucket sizes in multiples of two, we just need a right
bitmask to get it on the correct scale.

More details on this strategy here:
https://probablydance.com/2018/06/16/fibonacci-hashing-the-optimization-that-the-world-forgot-or-a-better-alternative-to-integer-modulo/