Lessen the number of spesh threshold levels
When something tiny called something a bit bigger, we could end up
with the call there not being able to specialize into a fast resolve
due to the callee potentially not being hot enough yet. Once things
get over a certain size that'll be noise, but there's plenty of cases
(tiny forwarding subs/methods delegating to slightly bigger impls)
that could benefit. So, collapse the lower levels to a single one.