Make parse_simple_number impossible to reach end without returning
Make sure that it's not possible to reach the end of the function
without returning or throwing. This fixes a compiler warning about this.

Also change the comments from // style to /* */ which is MoarVM's
recommended style.

Make an if else chain into a switch for greater clairity.

Remove some else's that were unneeded.