Fact discovery determines usage counts
Some time ago, a call was added to do fact discovery on inlinees. This
process sets usage counts along the way. However, the code to set the
usage counts in inlining was also in place, so we doubled all of the
counts, which would frustrate dead code elimination possibilities that
arise post-inline.

With this change we only do fact discovery, and that call is moved to
inline.c instead, where it feels a little more natural. Also don't
pass along the inliner's plan to go with inlinee fact discovery; it
doesn't relate to the inlinee, and could cause bogus guard insertion
if we're unlucky.