Attempt inlining so-far unspecialized callees
Sometimes we know the callee, but there wasn't yet a specialization
(or an applicable specialization) performed. If the bytecode is small
enough, then we can produce a specialization on demand and try to do
an inlining of that.

This doesn't yet convey the types of the incoming arguments to the
specialization of the callee, so various possible optimizations are
still not performed yet.