Add a maycausedeopt annotation
We'll use this to know when there's actual risk of deopt in the code
after we've optimzied it, but before we go doing dead instruction
elimination. Often, many potentially deopt-causing things go away due
to the devirtualization, and some log guards are simply unused. Doing
a post-optimization analysis will thus be more precise, and this new
annotation provides the tools to drive it.