Hash: randomize insertion order and bucket iteration order
The effect of this commit results in every hash having a different
iteration order. Previously, if you created two hashes and added items
in the same order, you would get the same iteration order.

Now we hash the table's pointer and use this as a pseudo random value to
choose iteration order. For a n bucket hash we XOR the top log2(n) bits
with the "bucket number". So we still use the same variable going from
0..7 for a table with 8 buckets, but if we XOR that with a 3 bit random
number for each bucket we get the numbers 0 through 7 but in a
pseudo random order.

We also randomize insertion order within a bucket. If there is already a
key in our chosen bucket, we randomly insert either as the first key or
as the second key.

The goal of these changes is to give an attacker less information about
the underlying hash values for each string. Without this change an
attacker could use iteration order as a great supplement to timing
attacks and can easily determine which keys end up in the same bucket.