Store deopt usage separate from normal usage
So that we can understand which things are kept alive for deopt and
which are being used for real. There's no behavior change from this,
it's simply making it clearer to see what lives only for the sake of
deopt in the spesh log. However, keeping this information separately
will allow us to do some further opts in the future.