Stop cheating on SSA form in optimize_call
Use a temporary register for now, which we already have code in place
to handle anyway becuase it's exactly what we do for the guard path.
Plus this should mean we can delete an unrequired earlier instruction
that's also reading from a spesh slot, to save some cycles. Most
importantly, however, it means we don't violate SSA form.