Make a faster HASH_ITER macro and make other hash optimizations
HASH_ITER_FAST is faster and it doesn't randomize bucket order (though
it's faster regardless of that).

Makes the biggest difference when we are iterating a hash with a LOT of
keys in it, such as adding a million keys to a hash then deleting and
adding back again. Got about 40% speed improvement on this nqp script:
https://gist.github.com/9a65a586ea21b20def8efe4aaaab55de
Most of the time spent in the above script was in MVMHash_gc_mark, and
that now uses the fast iterator internally which results in the speed
improvement.

Also optimize GET_X_BITS by making a switch out of it. It doesn't seem
like this would be faster, but it is due to compiler optimizations.