Correctly update OSR deopt index upon inlining
Otherwise, we end up with the wrong index, which leads to overwriting
the correct deopt offset with a bogus one. This caused a crash when
building NQP with JIT disabled, introduced since we started being able
to inline non-specialized code. This wasn't originally noticed, since
the JIT doesn't care for those indexes, and so everything worked under
JIT, and only failed in the case of running quickened bytecode.