Decrease instructions kept for potential deopt
We have to retain various instructions that are otherwise unused for
the sake of deopt. Eliminating these is rather delicate in general.
However, it's certain that any linear sequence of non-deopting
instructions at the end of the graph cannot possibly be impacted by
a deoptimization, since there is none that can take place. Thus, walk
backwards over those instructions, clearing their "required for deopt"
markers.