A new algorithm for removing unrequired `set`s
Checks if either the `set` is the only reader of the value that it
writes, or if there is only one user of the value that it writes.
Uses the DU-chain to quickly find the paired writer or reader, and
if there's on conflicting register use then does a graph rewrite to
enable deletion of the `set` instruction.

These seem to cover the cases that the previous more adhoc `set`
removal optimizations did, while also not making a mess of the SSA
form in the graph.