Liberalize set elimination somewhat
Sometimes we've no interesting control flow to worry about within the
graph: even if there may be calls to things outside the graph, those
can never touch our locals. Thus a linear sequence of instructions is
enough to look for to see we have on conflicts. Further, due to SSA
form, we need only look for *writes* that make a new version of the
potentially conflicting register; reads are uninteresting.

This generalization is in preparation for using this conflict
checking in order to help us do the non-temp case of box/unbox
elimination.