Implement box/unbox elimination
This happens in the second phase, after inlinings, thus giving us more
potential pairings to work with. When we see a box, we look at its
usages and find matching unbox/decont instructions. We also look down
`set` chains in case the unbox/decont is a `set` away. If we see a pair,
and the original unboxed value is still available, we rewrite the unbox
to a `set` of the original value, removing a usage of the boxed value.