Sketch out more precise deopt algorithm
As described in the commit body. It seems to do about the right thing,
except in the case of loops, where it makes a gross overestimate of
what needs to be preserved due to that part not yet being implemented.
Includes dumping of what deopt indices are keeping instructions alive.