Speed up hash garbage collection a lot (gc_mark); add new macros
Add new MVM_gc_worklist_add macros. These new macros assume that the
worklist is large enough to hold the new entries so MVM_gc_worklist_presize_for
must be called before hand removing one check and a potential function
call and hence why these are named `nocheck`.

There are two of these macros, one which assumes worklist->include_gen2
is true and one which assumes it is false. Since worklist->include_gen2
does not change in the use of a worklist, we can remove all checks in
the case include_gen2 is true, and remove all but one in the case it is
true.

Additionally they assume that the item to be added is non-null, so
should only be used when GCing things which have been checked for being
non-null. Here we use it for hashes which are checked they aren't null
before being added to the hash.

This improves the speed of MVMHash_gcmark by at least 20%. And improves
the speed of this nqp script by 10%
https://gist.github.com/samcv/9a65a586ea21b20def8efe4aaaab55de#file-time-nqp