Switch over to using the new deopt use method
We don't currently every throw away any deopt usages, so in the
immediate this is a step backwards (since it removes the "can't do
a deopt after the last deopting instruction" optimization). However,
this will let us work out places where we fail to keep something alive
for deopt that we should be.