Don't repeatedly calcuate if preds are seen
We can instead keep track of which basic blocks have had all of their
preds handled, and update it by looking at the succs of a basic block
after processing it. Better still, if nothing changes, then we know we
don't have to both processing the outstanding reads at all.