Move speshresolve handling into optimize phase
Previously, the guards were inserted at the facts phase. We got away
with this when deopt processing was more lax and in band with the facts
analysis. However, now it takes place after. That's still OK for most
guard insertion cases. But spesh plugins aren't a guard insertion. They
are a significant graph change, rendering the graph too different from
what we started out with for the deopt analysis pass to produce accurate
results. That could lead to very occasional crashes on deopt due to some
instructions being deleted that should not have been.

Given the code is doing something very much like what optimize_call does
in the case it looks at the logs, it is arguably more natural to do this
in the optimize phase anyway. Plus we might have found out more, and so
have less guards to insert. On one hand, we'll be building redundant
guard elimination soon anyway, on the other it's nice to not insert
something only to throw it away again later.