Preserve enough data for accurate traversal
In the absence of inlining, the ->caller chain is immutable, so we can
simply hold the frame the `ctx` op ran on and it's enough. When there's
inlining, however, we didn't have sufficient information in order to
be able to reconstruct the caller chain, since the current address of
the caller is mutable. Thus, snapshot those at the point of a use of
the `ctx` op in the *callee*. A callee can only ever have been called
from a particular program point, and so this lets us better traverse
the chain.