Fix delete motion of INLINE_END annotations
We should always move them forwards. We never had a problem with this
in the past, since the last instruction of an inline was always a
`set` that was replacing the `return`. That was never deleted, so we
never had a problem. Now we support inlining of `invoke_v`, which means
the return instruction is simply deleted. When we moved the inline end
annotation back one, we could move it into an invoke instruction. This
in turn meant that the deopt all index of that invoke would not fall
within the bounds of the inline any more, and so we would deopt to an
index intended for the most nested inlinee but apply it to the code
outside of that. Effectively, this placed the interpreter back in an
utterly bogus position, and pretty much anything could happen.