Don't use 32-bit unaligned reads in SipHash implementation
This fixes a bug on CentOS 6 and possibly is a bug in glibc 2.12.1 or
the toolchain which has issues with unaligned 32-bit reads.

Instead of trying to do a 32-bit unaligned read, only do 8-bit reads.
This should optimize to a single operation in the compiler, and even if
it does not, we only called siphashfinish or siphashfinish_32bits once
for every string we hash.

Fixes issue #910