Make sure output of spesh resolve gets optimized, too
Until now we'd have resumed optimization after the last
instruction a speshresolve has put in. This misses the
opportunity to improve getattr faux-guards into direct
pointer access versions.

Fix this by remembering the previous instruction before
speshresolve is optimized and continuing with that
instruction. The next thing in the loop is to go to
ins->next, so we use prev->prev.