Use better solution to SipHash fix for GCC 4
The previous solution was slower than I would have hoped. Although it
worked, using unions reduce the amount of type punning we need to do. It
is also faster than the version before the fix was introduced while not
triggering the compiler issue in GCC 4. It is also more elegant and nice
to look at.