Add "standard" siphash tests using siphashadd64bits + large time test
Also test siphashadd64bits with the same tests that we know for sure to
be correct (the ones that come from the original upstream siphash
source). We know these are right so add further coverage. This helps us
know more fully that our MVM specific tests are correct.

Also add a test meant to be used for timing very long strings (10,000
graphemes in this case).