Make spesh thread more GC-responsive
By making us able to mark the spesh graph more completely, and making
the one we're currently working on available for GC marking, we can
introduce GC sync points at a number of locations in the optimization
process. This can notably reduce the GC latency when optimizing larger
graphs.

Didn't measure repeatedly, but spectest time decreased by ~8s and
CORE.setting compilation by ~2s; the latter feels a bit on the high
side, so I'd take these numbers with a pinch of salt.