[Spesh] Leak fixes
Many allocated things either come from a spesh candidate or from
malloc (or realloc). Or they are assigned to a spesh candidate.
So the only way to really know if they are shared with a spesh candidate
(or static frame) is by checking if they are pointer-equal.

ASAN leaksanitizer is finally clean now for me.