Don't do worthless type specializations
Note whether a static frame contains anything that would benefit from
type specialization. If it contains no such instructions, then only do
a certain specialization of it. This works well with the `identity` sub
written in NQP and produced by various spesh plugins. With a bit more
work in Rakudo it may be possible to get Mu.sink to be considered as an
unspecializable thing too.