Do dead instruction elim pass after usage setup
But before we do the deopt analysis. While it's of course possible
that there was dead code in the input program, the main purpose of
this is to trim away dead PHI nodes, recursively, so that they will
not make extra work for the deopt analyzer nor cause things to be
kept alive bogusly just because a PHI was used across a deopt point.

This seems to have a positive effect on spectest time, probably due
to the deopt usage algorithm having a bit less to do, as well as
resulting in elimination of more instructions during optimization
that were falsely retained.